Faas 6c: tõsta collections ja archives routerisse #36
